backend/requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
cffi==1.16.0
channels==4.0.0
constantly==23.10.4
cryptography==41.0.7
daphne==4.0.0
Django==5.0
djangorestframework==3.14.0
frozenlist==1.4.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
livekit-api==0.3.0
livekit-protocol==0.2.0
multidict==6.0.4
Pillow==10.1.0
protobuf==4.25.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==23.3.0
pytz==2023.3.post1
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.4
Twisted==23.10.0
txaio==23.1.1
types-protobuf==4.24.0.4
typing_extensions==4.9.0
yarl==1.9.4
zope.interface==6.1
